get rid of MUI obsolete Grid component
