+++ next/app/admin/room/[ruid]/bans/page.tsx
@@ -5,8 +5,7 @@ import {
   Container, Grid2 as Grid, Paper, Table, TableBody,
   TableCell, TableHead, TableRow, Button, Divider, 
   IconButton, TextField, Typography, 
-  useTheme,
-  Theme
+  useTheme, Theme
 } from '@mui/material';
 import WidgetTitle from '@/components/Admin/WidgetTitle';
 import client from '@/lib/client';
