+++ next/app/admin/room/[ruid]/page.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import React, { useContext, useEffect, useState } from 'react';
-import clsx from 'clsx';
 import { 
   Container, Grid2 as Grid, Paper, Button, 
   Divider, TextField, Alert,
