+++ next/components/SignIn.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState } from 'react';
 import { 
-  Avatar, Button, CssBaseline, TextField, Box,
+  Avatar, Button, CssBaseline, TextField,
   Typography, Container, styled 
 } from '@mui/material';
 import { LockOutlined } from '@mui/icons-material';
